最后几回合降低激进度

+++ UmaSimulator/GameDatabase/GameConfig.h
@@ -10,7 +10,7 @@
 struct GameConfig 
 {
     static bool noColor;    // 为True时不显示颜色
-    static int radicalFactor; // 激进度，默认为5，提高会导致计算变慢
+    static double radicalFactor; // 激进度，默认为5，提高会导致计算变慢
     static int threadNum;   // 线程数，默认为12，可根据自身CPU调整
     static int searchN;  // 迭代轮数，默认为6144，缩小该值可以加快速度，但会降低准确度
     static bool debugPrint; // 显示调试信息，例如进度条'.'，默认为False
